refactor(CalculatorHelper) : rename CalculatorUtils
